// src/app/models/provider.ts
export interface Provider {
    id: string;
    name: string;
    phone: string;
    contact: string;
    street: string;
    num: string;
    colony: string;
    pc: number;
    city: string;
    state: string;
    cellphone: string;
    email: string;
    bank: string;
    account: string;
    clabe: string;
    category: string;
}